Test validating zip upload and extraction.
